Quote cookie values
According to [RFC2019](https://tools.ietf.org/html/rfc2109), this is
required for values containing spaces, and should always be safe for
non-space containing values.

Note that most webservers are quite smart and lenient when parsing
cookie headers. This change however is still useful for servers that
are very strict (such as undertow).