Add a common base class for requests exceptions
Sometimes, one wants to catch and handle all exceptions thrown
by requests regardless if it is an HTTP 500 due to a server problem
or a DNS problem. This patch adds such a base class which makes
error handling easier.

This class is similar in purpose to RequestException in the Python
requests library.